+++ bibliotheque-api/bibliotheque-api-consumer/src/main/java/com/bibliotheque/api/consumer/ReservationRepository.java
@@ -18,7 +18,7 @@ public interface ReservationRepository extends JpaRepository<Reservation, Long> 
             "WHERE (:livre IS NULL OR r.livre = :livre) \n" +
             "AND (:utilisateur IS NULL OR r.utilisateur = :utilisateur )" +
             " \n" +
-            "AND r.dateDebut <= :date" +
+            "AND r.dateCreation <= :date" +
     " AND r.rendu = false ")
     List<Reservation> findActualReservationsWithLivre(@Param("livre") Livre livre, @Param("utilisateur") Utilisateur utilisateur, @Param("date") DateTime date);
