+++ bibliotheque-webapp/src/main/java/com/bibliotheque/webapp/WebappApplication.java
@@ -27,6 +27,7 @@ public class WebappApplication {
         UtilisateurApi serviceUtilisateur = retrofit.create(UtilisateurApi.class);
         String base = "fjoseph24@orange.fr:azerty123";
 */
+
         SpringApplication.run(WebappApplication.class, args);
     }
 }
