+++ bibliotheque-webapp/src/main/resources/templates/livre.html
@@ -19,10 +19,18 @@
                     <h2 class="subtitle is-4 is-thin"
                         th:text="${livre.getAuteur()} + ', ' + ${livre.getAnnee()} + ', ' + ${livre.getGenre()}"></h2>
                     <!-- CTA -->
-                    <h2 class="subtitle is-8 is-thin" th:text="'Livres disponibles : ' + ${reservationsSize}"></h2>
-                    <h2 class="subtitle is-8 is-thin" th:text="'Réservations en attente : ' + ${reservationsAttenteSize}"></h2>
+                    <h2 class="subtitle is-8 is-thin" th:text="'Livres disponibles : ' + ${livresDispo}"></h2>
+                    <h2 th:if="${livresDispo == 0}" class="subtitle is-8 is-thin"
+                        th:text="'Réservations en attente : ' + ${reservationsAttente}"></h2>
+                    <h2 th:if="${livresDispo == 0}" class="subtitle is-8 is-thin" th:text="'Date retour : ' +
+                    ${dateRetour.getYear()} +
+                    '/' +
+                    ${dateRetour.getMonthOfYear()} +
+                    '/' +
+                    ${dateRetour.getDayOfMonth()}"></h2>
                     <p>
-                        <a th:href="@{/reservation/{id} (id=${livre.getId()})}" class="button is-rounded is-large custom-background has-text-white">
+                        <a th:href="@{/reservation/{id} (id=${livre.getId()})}"
+                           class="button is-rounded is-large custom-background has-text-white">
                             <span class="text">Réserver</span>
                             <span class="front-gradient"></span>
                         </a>
